requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
